backend/pyproject.toml
[project]
name = "ncu-tldr-backend"
version = "0.1.0.dev1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0",
    "email-validator>=2.3.0",
    "fastapi>=0.118.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.15.2",
]
